Move JSDescription::Verify to ProofVerifier::VerifySprout
