+++ src/primitives/transaction.h
@@ -7,7 +7,6 @@
 #define BITCOIN_PRIMITIVES_TRANSACTION_H
 
 #include "amount.h"
-#include "proof_verifier.h"
 #include "random.h"
 #include "script/script.h"
 #include "serialize.h"
@@ -259,12 +258,6 @@ public:
             uint256 *esk = nullptr, // payment disclosure
             std::function<int(int)> gen = GetRandInt
     );
-
-    // Verifies that the JoinSplit proof is correct.
-    bool Verify(
-        ProofVerifier& verifier,
-        const uint256& joinSplitPubKey
-    ) const;
 
     // Returns the calculated h_sig
     uint256 h_sig(const uint256& joinSplitPubKey) const;
