+++ src/wallet/asyncrpcoperation_shieldcoinbase.cpp
@@ -320,7 +320,7 @@ UniValue AsyncRPCOperation_shieldcoinbase::perform_joinsplit(ShieldCoinbaseJSInf
             &esk); // parameter expects pointer to esk, so pass in address
     {
         auto verifier = ProofVerifier::Strict();
-        if (!(jsdesc.Verify(verifier, joinSplitPubKey_))) {
+        if (!(verifier.VerifySprout(jsdesc, joinSplitPubKey_))) {
             throw std::runtime_error("error verifying joinsplit");
         }
     }
